+++ out/extension.js
@@ -11,15 +11,7 @@ function activate(context) {
     // Use the console to output diagnostic information (console.log) and errors (console.error)
     // This line of code will only be executed once when your extension is activated
     console.log('Congratulations, your extension "flutter-project-template" is now active!');
-    // The command has been defined in the package.json file
-    // Now provide the implementation of the command with registerCommand
-    // The commandId parameter must match the command field in package.json
-    let disposable = vscode_1.commands.registerCommand('flutter-project-template.helloWorld', () => {
-        // The code you place here will be executed every time your command is executed
-        // Display a message box to the user
-        vscode_1.window.showInformationMessage('Hello World from Flutter Project Template!');
-    });
-    context.subscriptions.push(vscode_1.commands.registerCommand('flutter-project-template.setup', commands_1.setup));
+    context.subscriptions.push(vscode_1.commands.registerCommand('flutter-project-template.setup', commands_1.setup), vscode_1.commands.registerCommand('flutter-project-template.new-screen', commands_1.newScreen));
 }
 exports.activate = activate;
 // This method is called when your extension is deactivated