+++ out/utils/file-utils.js
@@ -1,6 +1,6 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.createDirectory = exports.createFile = void 0;
+exports.addToFile = exports.createDirectory = exports.createFile = void 0;
 const fs_1 = require("fs");
 const mkdirp = require("mkdirp");
 function createFile(filePath, content) {
@@ -26,4 +26,16 @@ function createDirectory(targetDirectory) {
     });
 }
 exports.createDirectory = createDirectory;
+function addToFile(filePath, contentToAdd) {
+    return new Promise(async (resolve, reject) => {
+        (0, fs_1.appendFile)(filePath, contentToAdd, "utf8", (error) => {
+            if (error) {
+                reject(error);
+                return;
+            }
+            resolve();
+        });
+    });
+}
+exports.addToFile = addToFile;
 //# sourceMappingURL=file-utils.js.map