+++ src/commands/index.ts
@@ -1,1 +1,2 @@
 export * from "./setup.command";
+export * from "./new-screen.command";