+++ src/extension.ts
@@ -1,7 +1,7 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 import { commands, ExtensionContext, window, languages, workspace } from "vscode";
-import { setup } from "./commands";
+import { setup, newScreen } from "./commands";
 
 // This method is called when your extension is activated
 // Your extension is activated the very first time the command is executed
@@ -11,17 +11,9 @@ export function activate(context: ExtensionContext) {
 	// This line of code will only be executed once when your extension is activated
 	console.log('Congratulations, your extension "flutter-project-template" is now active!');
 
-	// The command has been defined in the package.json file
-	// Now provide the implementation of the command with registerCommand
-	// The commandId parameter must match the command field in package.json
-	let disposable = commands.registerCommand('flutter-project-template.helloWorld', () => {
-		// The code you place here will be executed every time your command is executed
-		// Display a message box to the user
-		window.showInformationMessage('Hello World from Flutter Project Template!');
-	});
-
 	context.subscriptions.push(
     commands.registerCommand('flutter-project-template.setup', setup),
+    commands.registerCommand('flutter-project-template.new-screen', newScreen),
   );
 }