+++ src/utils/file-utils.ts
@@ -1,4 +1,4 @@
-import { writeFile } from "fs";
+import { writeFile, appendFile } from "fs";
 import * as mkdirp from "mkdirp";
 
 export function createFile(filePath: string, content: string): Promise<void>{
@@ -26,5 +26,22 @@ export function createDirectory(targetDirectory: string): Promise<void> {
       }
       resolve();
     });
+  });
+}
+
+export function addToFile(filePath: string, contentToAdd: string): Promise<void>{
+  return new Promise<void>(async (resolve, reject) => {
+    appendFile(
+      filePath,
+      contentToAdd,
+      "utf8",
+      (error) => {
+        if (error) {
+          reject(error);
+          return;
+        }
+        resolve();
+      }
+    );
   });
 }