+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to the "flutter-project-template" extension will be document
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [0.2.0] - 2023-01-04
+
+### Added
+
+- New Screen command
+- fpsimpleroute snippet
+
+## [0.1.2] - 2023-01-03
+
+### Fixed
+
+- Fixed dependencies
+
+## [0.1.0] - 2023-01-03
 
 - Initial release